services/chatbot/requirements.txt
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-mongodb==0.6.1
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
markdown==3.8
pymongo==4.12.1
python-dotenv==1.1.0
unstructured==0.17.2
numpy==1.26.4
langchain-mcp-adapters==0.1.8
Quart==0.20.0
quart-cors==0.8.0
motor==3.7.1
openai==1.77.0
langgraph==0.5.1
faiss-cpu==1.11.0
psycopg2-binary
uvicorn==0.35.0
fastmcp==2.10.2